Simplify compute test env (#92)
* Refine compute test env

* Revert more changed codef

* Revert more changed code

* Fix bridge test build

* Split arrow_compute_testing into two object libraries for tests that will link to libarrow_compute and tests that won't

* Add ARROW_GTEST_GTEST_MAIN to target_link_libraries too

* Use arrow_compute_testing instead of arrow_compute_core_testing for row_test

---------
