+++ cpp/src/arrow/acero/hash_aggregate_test.cc
@@ -39,7 +39,6 @@
 #include "arrow/compute/cast.h"
 #include "arrow/compute/exec.h"
 #include "arrow/compute/exec_internal.h"
-#include "arrow/compute/kernels/test_util_internal.h"
 #include "arrow/compute/registry.h"
 #include "arrow/compute/row/grouper.h"
 #include "arrow/table.h"
@@ -68,7 +67,6 @@ using internal::ToChars;
 
 using compute::ArgShape;
 using compute::CallFunction;
-using compute::ComputeKernelEnvironment;
 using compute::CountOptions;
 using compute::default_exec_context;
 using compute::ExecBatchFromJSON;
@@ -87,10 +85,6 @@ using compute::Take;
 using compute::TDigestOptions;
 using compute::ValidateOutput;
 using compute::VarianceOptions;
-
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
 
 namespace acero {
