+++ cpp/src/arrow/acero/plan_test.cc
@@ -27,7 +27,6 @@
 #include "arrow/acero/util.h"
 #include "arrow/compute/exec.h"
 #include "arrow/compute/expression.h"
-#include "arrow/compute/kernels/test_util_internal.h"
 #include "arrow/compute/test_util_internal.h"
 #include "arrow/io/util_internal.h"
 #include "arrow/record_batch.h"
@@ -55,7 +54,6 @@ namespace arrow {
 
 using compute::ArgShape;
 using compute::call;
-using compute::ComputeKernelEnvironment;
 using compute::CountOptions;
 using compute::ExecBatchFromJSON;
 using compute::field_ref;
@@ -64,10 +62,6 @@ using compute::SortKey;
 using compute::SortOrder;
 using compute::Take;
 using compute::TDigestOptions;
-
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
 
 namespace acero {
