+++ cpp/src/arrow/compute/kernels/aggregate_test.cc
@@ -56,10 +56,6 @@ namespace compute {
 
 using internal::FindAccumulatorType;
 
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
-
 //
 // Sum
 //
