+++ cpp/src/arrow/compute/kernels/scalar_arithmetic_test.cc
@@ -44,11 +44,6 @@
 
 namespace arrow {
 namespace compute {
-
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
-
 namespace {
 
 // 2.718281828459045090795598298427648842334747314453125
