+++ cpp/src/arrow/compute/kernels/scalar_if_else_test.cc
@@ -34,10 +34,6 @@ using internal::checked_pointer_cast;
 
 namespace compute {
 
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
-
 // Helper that combines a dictionary and the value type so it can
 // later be used with DictArrayFromJSON
 struct JsonDict {
