+++ cpp/src/arrow/dataset/file_ipc_test.cc
@@ -21,7 +21,6 @@
 #include <utility>
 #include <vector>
 
-#include "arrow/compute/kernels/test_util_internal.h"
 #include "arrow/dataset/dataset_internal.h"
 #include "arrow/dataset/discovery.h"
 #include "arrow/dataset/file_base.h"
@@ -38,12 +37,7 @@
 
 namespace arrow {
 
-using compute::ComputeKernelEnvironment;
 using internal::checked_pointer_cast;
-
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
 
 namespace dataset {
