+++ cpp/src/arrow/dataset/file_parquet_encryption_test.cc
@@ -21,7 +21,6 @@
 
 #include "arrow/array.h"
 #include "arrow/compute/api_vector.h"
-#include "arrow/compute/kernels/test_util_internal.h"
 #include "arrow/dataset/dataset.h"
 #include "arrow/dataset/file_base.h"
 #include "arrow/dataset/file_parquet.h"
@@ -53,13 +52,6 @@ constexpr std::string_view kBaseDir = "";
 using arrow::internal::checked_pointer_cast;
 
 namespace arrow {
-
-using compute::ComputeKernelEnvironment;
-
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
-
 namespace dataset {
 
 struct EncryptionTestParam {
